#85 move item count to bottom of screen
#84 adjust font sizes
fix height on properties page
add fallback for user logged in status
fix typo in readme
