+++ KeyVaultExplorer/Views/CustomControls/UserStatusTag.axaml
@@ -1,0 +1,67 @@
+﻿<UserControl
+    x:Class="KeyVaultExplorer.Views.CustomControls.UserStatusTag"
+    xmlns="https://github.com/avaloniaui"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:models="clr-namespace:KeyVaultExplorer.Models;assembly=KeyVaultExplorer"
+    xmlns:ui="using:FluentAvalonia.UI.Controls"
+    xmlns:vm="using:KeyVaultExplorer.ViewModels"
+    d:DesignHeight="150"
+    d:DesignWidth="925"
+    x:DataType="vm:MainViewModel"
+    mc:Ignorable="d">
+
+    <Design.DataContext>
+        <vm:MainViewModel />
+    </Design.DataContext>
+
+
+    <StackPanel
+        Grid.Row="1"
+        HorizontalAlignment="Stretch"
+        Background="{DynamicResource SolidBackgroundFillColorTertiaryBrush}"
+        IsHitTestVisible="False">
+
+        <StackPanel
+            HorizontalAlignment="Right"
+            IsVisible="{Binding !IsAuthenticated}"
+            Orientation="Horizontal">
+            <ui:BitmapIcon
+                Width="16"
+                Height="16"
+                VerticalAlignment="Bottom"
+                RenderOptions.BitmapInterpolationMode="HighQuality"
+                UriSource="avares://KeyVaultExplorer/Assets/StatusErrorOutline.png" />
+            <TextBlock
+                Margin="4,0"
+                Background="{x:Null}"
+                FontSize="{StaticResource FontSizeSmall}"
+                Text="You are currently signed out"
+                Theme="{StaticResource CaptionTextBlockStyle}" />
+        </StackPanel>
+
+
+        <StackPanel
+            HorizontalAlignment="Right"
+            IsVisible="{Binding IsAuthenticated}"
+            Orientation="Horizontal">
+            <ui:BitmapIcon
+                Width="16"
+                Height="16"
+                Margin="0,2,0,0"
+                VerticalAlignment="Bottom"
+                IsVisible="{Binding AuthenticatedUserClaims.TenantId}"
+                RenderOptions.BitmapInterpolationMode="HighQuality"
+                ShowAsMonochrome="True"
+                UriSource="avares://KeyVaultExplorer/Assets/Cloud.png" />
+            <TextBlock
+                Margin="5,0"
+                Background="{x:Null}"
+                FontSize="{StaticResource FontSizeSmall}"
+                Text="{Binding AuthenticatedUserClaims.Email}"
+                Theme="{StaticResource CaptionTextBlockStyle}" />
+        </StackPanel>
+    </StackPanel>
+
+</UserControl>
