+++ KeyVaultExplorer/Views/MainPage.axaml
@@ -2,6 +2,7 @@
     x:Class="KeyVaultExplorer.Views.Pages.MainPage"
     xmlns="https://github.com/avaloniaui"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:custom="clr-namespace:KeyVaultExplorer.Views.CustomControls"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:models="clr-namespace:KeyVaultExplorer.Models;assembly=KeyVaultExplorer"
@@ -17,61 +18,7 @@
         <vm:MainViewModel />
     </Design.DataContext>
 
+    <pages:TabViewPage Name="TabViewPage" Background="{DynamicResource LayerOnMicaBaseAltFillColorSecondaryBrush}" />
 
-    <Grid Name="MainGrid" RowDefinitions="*, 18">
-        <pages:TabViewPage
-            Name="TabViewPage"
-            Grid.Row="0"
-            Background="{DynamicResource LayerOnMicaBaseAltFillColorSecondaryBrush}" />
-        <StackPanel
-            Grid.Row="1"
-            HorizontalAlignment="Stretch"
-            Background="{DynamicResource SolidBackgroundFillColorTertiaryBrush}"
-            IsHitTestVisible="False">
-
-            <StackPanel
-                HorizontalAlignment="Right"
-                IsVisible="{Binding !IsAuthenticated}"
-                Orientation="Horizontal">
-                <ui:BitmapIcon
-                    Width="16"
-                    Height="16"
-                    VerticalAlignment="Bottom"
-                    RenderOptions.BitmapInterpolationMode="HighQuality"
-                    UriSource="avares://KeyVaultExplorer/Assets/StatusErrorOutline.png" />
-                <TextBlock
-                    Margin="4,0"
-                    Background="{x:Null}"
-                    FontSize="{StaticResource FontSizeSmall}"
-                    Text="You are currently signed out." />
-                <!--    <TextBlock
-                    Margin="4,0"
-                    Background="{x:Null}"
-                    FontSize="{StaticResource FontSizeSmall}"
-                    Text="Sign In"
-                    TextDecorations="Underline" />-->
-            </StackPanel>
-
-
-            <StackPanel
-                HorizontalAlignment="Right"
-                IsVisible="{Binding IsAuthenticated}"
-                Orientation="Horizontal">
-                <ui:BitmapIcon
-                    Width="16"
-                    Height="16"
-                    VerticalAlignment="Bottom"
-                    IsVisible="{Binding AuthenticatedUserClaims.TenantId}"
-                    RenderOptions.BitmapInterpolationMode="HighQuality"
-                    ShowAsMonochrome="True"
-                    UriSource="avares://KeyVaultExplorer/Assets/Cloud.png" />
-                <TextBlock
-                    Margin="5,0"
-                    Background="{x:Null}"
-                    FontSize="{StaticResource FontSizeSmall}"
-                    Text="{Binding AuthenticatedUserClaims.Email}" />
-            </StackPanel>
-        </StackPanel>
-    </Grid>
 
 </UserControl>
